Clean, correct, refactor

+++ Server/app/controllers/GroupController.java
@@ -71,9 +71,12 @@ public class GroupController extends Controller {
             return forbidden("No requester specified");
         }
         Integer requesterId = Integer.parseInt(request().getHeader("requester"));
-        System.out.println("REQESTOR:" +  requesterId);
         Session session = SessionHandler.getInstance().getSessionFactory().openSession();
         Group group = session.get(Group.class, groupid);
+        if(group == null) {
+            session.close();
+            return notFound();
+        }
         if (group.getOwner() == requesterId) {
             session.getTransaction().begin();
             session.delete(group);
